+++ flutter_inappwebview/example/lib/in_app_webiew_example.screen.dart
@@ -24,6 +24,7 @@ class _InAppWebViewExampleScreenState extends State<InAppWebViewExampleScreen> {
       javaScriptHandlerOriginAllowList: {".*"},
       pluginScriptsForMainFrameOnly: false,
       pluginScriptsOriginAllowList: {"*"},
+      isFindInteractionEnabled: false,
       iframeAllowFullscreen: true);
 
   PullToRefreshController? pullToRefreshController;
@@ -129,14 +130,16 @@ class _InAppWebViewExampleScreenState extends State<InAppWebViewExampleScreen> {
                   pullToRefreshController: pullToRefreshController,
                   initialUserScripts: UnmodifiableListView<UserScript>([
                     UserScript(
-                        source: "console.log('loaded'); window.custom_js_bridge.callHandler('handlerName', 1, 4, true, {\"foo\": \"bar\"});",
+                        source: """console.log('loaded', window.location.href); window.custom_js_bridge.callHandler('handlerName', 1, 4, true, {'foo': 'bar'});""",
                         injectionTime: UserScriptInjectionTime.AT_DOCUMENT_END,
-                        allowedOriginRules: {"https://www.w3schools.com", "https://www.example.com"},
-                        forMainFrameOnly: false),
+                        forMainFrameOnly: true),
                   ]),
                   onWebViewCreated: (controller) async {
                     webViewController = controller;
                     controller.addJavaScriptHandler(handlerName: 'handlerName', callback: (JavaScriptHandlerFunctionData handlerData) {
+                      if (handlerData.origin.host != "www.w3schools.com" || !handlerData.isMainFrame) {
+                        throw Exception("This is an exception from a JavaScript handler");
+                      }
                       print(handlerData);
                       return handlerData.args;
                     });