+++ flutter_inappwebview_ios/ios/Classes/PluginScriptsJS/CallAsyncJavaScriptBelowIOS14WrapperJS.swift
@@ -7,15 +7,28 @@
 
 import Foundation
 
-let CALL_ASYNC_JAVASCRIPT_BELOW_IOS_14_WRAPPER_JS = """
-(function(obj) {
-    (async function(\(PluginScriptsUtil.VAR_FUNCTION_ARGUMENT_NAMES) {
-        \(PluginScriptsUtil.VAR_FUNCTION_BODY)
-    })(\(PluginScriptsUtil.VAR_FUNCTION_ARGUMENT_VALUES)).then(function(value) {
-        window.webkit.messageHandlers['onCallAsyncJavaScriptResultBelowIOS14Received'].postMessage({'value': value, 'error': null, 'resultUuid': '\(PluginScriptsUtil.VAR_RESULT_UUID)'});
-    }).catch(function(error) {
-        window.webkit.messageHandlers['onCallAsyncJavaScriptResultBelowIOS14Received'].postMessage({'value': null, 'error': error + '', 'resultUuid': '\(PluginScriptsUtil.VAR_RESULT_UUID)'});
-    });
-    return null;
-})(\(PluginScriptsUtil.VAR_FUNCTION_ARGUMENTS_OBJ));
-"""
+public class CallAsyncJavaScriptBelowIOS14WrapperJS {
+    
+    public static func CALL_ASYNC_JAVASCRIPT_BELOW_IOS_14_WRAPPER_JS() -> String {
+        return """
+        (function(obj) {
+            (async function(\(PluginScriptsUtil.VAR_FUNCTION_ARGUMENT_NAMES) {
+                \(PluginScriptsUtil.VAR_FUNCTION_BODY)
+            })(\(PluginScriptsUtil.VAR_FUNCTION_ARGUMENT_VALUES)).then(function(value) {
+                window.\(JavaScriptBridgeJS.get_JAVASCRIPT_BRIDGE_NAME()).callHandler('onCallAsyncJavaScriptResultBelowIOS14Received', {
+                    'value': value, 
+                    'error': null,
+                    'resultUuid': '\(PluginScriptsUtil.VAR_RESULT_UUID)'
+                });
+            }).catch(function(error) {
+                window.\(JavaScriptBridgeJS.get_JAVASCRIPT_BRIDGE_NAME()).callHandler('onCallAsyncJavaScriptResultBelowIOS14Received', {
+                    'value': null, 
+                    'error': error + '',
+                    'resultUuid': '\(PluginScriptsUtil.VAR_RESULT_UUID)'
+                });
+            });
+            return null;
+        })(\(PluginScriptsUtil.VAR_FUNCTION_ARGUMENTS_OBJ));
+        """
+    }
+}