+++ flutter_inappwebview_ios/ios/Classes/PluginScriptsJS/WebMessageChannelJS.swift
@@ -7,4 +7,10 @@
 
 import Foundation
 
-let WEB_MESSAGE_CHANNELS_VARIABLE_NAME = "window.\(JAVASCRIPT_BRIDGE_NAME)._webMessageChannels"
+public class WebMessageChannelJS {
+    
+    public static func WEB_MESSAGE_CHANNELS_VARIABLE_NAME() -> String {
+        return "window.\(JavaScriptBridgeJS.get_JAVASCRIPT_BRIDGE_NAME())._webMessageChannels"
+    }
+    
+}