+++ flutter_inappwebview_platform_interface/lib/src/types/user_script.g.dart
@@ -10,8 +10,15 @@ part of 'user_script.dart';
 class UserScript {
   ///A set of matching rules for the allowed origins.
   ///
+  ///**NOTE for Android**: each origin pattern MUST follow the table rule of [PlatformInAppWebViewController.addWebMessageListener].
+  ///
+  ///**NOTE for iOS and macOS**: each origin pattern will be used as a
+  ///Regular Expression Pattern that will be used on JavaScript side using [RegExp](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp).
+  ///
   ///**Officially Supported Platforms/Implementations**:
   ///- Android native WebView
+  ///- iOS
+  ///- MacOS
   late Set<String> allowedOriginRules;
 
   ///A scope of execution in which to evaluate the script to prevent conflicts between different scripts.